code factorisation
